requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
build==1.2.2
    # via pip-tools
click==8.1.7
    # via pip-tools
imagehash==4.3.1
    # via -r requirements.in
numpy==2.1.1
    # via
    #   imagehash
    #   pywavelets
    #   scipy
packaging==24.1
    # via build
pillow==10.4.0
    # via imagehash
pip-tools==7.4.1
    # via -r requirements.in
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pywavelets==1.7.0
    # via imagehash
scipy==1.14.1
    # via imagehash
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
